create new file api.js
