+++ src/api/api-popularlist.js
@@ -1,22 +1,5 @@
-import axios from "axios";
 import { baseUrl } from "./base";
-
-const axiosInstance = axios.create({
-    baseURL: baseUrl,
-  });
-  
-  axiosInstance.interceptors.request.use(
-    (config) => {
-      const token = localStorage.getItem('token');
-      if (token) {
-        config.headers['Authorization'] = `Bearer ${token}`;
-      }
-      return config;
-    },
-    (error) => {
-      console.error(error);
-    },
-  );
+import { axiosInstance } from "./api.js"
 
   export const getPopularList = async() => {
     try{
