+++ src/api/api-replylist.js
@@ -1,0 +1,13 @@
+import { baseUrl } from "./base";
+import { axiosInstance } from "./api.js"
+
+export 
+const getReplyList = async () => {
+  try{
+
+  }catch(error){
+      console.error("Get replyList failed")
+  }
+}
+
+
